Added exception logic for products.

+++ src/main/java/com/springbootmybatis/service/impl/ProductServiceImpl.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.springbootmybatis.entity.Product;
 import com.springbootmybatis.mapper.ProductMapper;
 import com.springbootmybatis.service.IProductService;
+import org.apache.ibatis.javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,8 +18,14 @@ public class ProductServiceImpl extends ServiceImpl<ProductMapper, Product> impl
     ProductMapper productMapper;
 
     @Override
-    public Product getProductById(Integer productId) {
-        return productMapper.selectById(productId);
+    public Product getProductById(Integer productId) throws NotFoundException {
+        Product product = productMapper.selectById(productId);
+        if (product != null) {
+            return product;
+        } else {
+            log.error("Error in method getProductById! Product with id '" + productId + "' not found!");
+            throw new NotFoundException("Product with id '" + productId + "' not found!");
+        }
     }
 
     @Override
@@ -39,16 +46,35 @@ public class ProductServiceImpl extends ServiceImpl<ProductMapper, Product> impl
     }
 
     @Override
-    public void deleteProduct(Integer productId) {
-        productMapper.deleteById(productId);
+    public void deleteProduct(Integer productId) throws NotFoundException {
+        Product product = productMapper.selectById(productId);
+        if (product != null) {
+            productMapper.deleteById(productId);
+        } else {
+            log.error("Error in method getProductById! Product with id '" + productId + "' not found!");
+            throw new NotFoundException("Product with id '" + productId + "' not found!");
+        }
+
     }
 
     @Override
-    public Product updateProduct(Integer productId, Product product) {
-        QueryWrapper<Product> queryWrapper = new QueryWrapper<>();
-        queryWrapper.eq("product_id", productId);
-        productMapper.update(product, queryWrapper);
-        return product;
+    public Product updateProduct(Integer productId, Product product) throws NotFoundException {
+        Product productFromDb = productMapper.selectById(productId);
+        if (productFromDb != null) {
+            productFromDb.setProducerId(product.getProducerId());
+            productFromDb.setName(product.getName());
+            productFromDb.setCategory(product.getCategory());
+            productFromDb.setPrice(product.getPrice());
+            productFromDb.setCreatedOn(product.getCreatedOn());
+            productFromDb.setExpirationDate(product.getExpirationDate());
+            QueryWrapper<Product> queryWrapper = new QueryWrapper<>();
+            queryWrapper.eq("product_id", productId);
+            productMapper.update(productFromDb, queryWrapper);
+            return productFromDb;
+        } else {
+            log.error("Error in method updateProduct! Product with id '" + productId + "' not found!");
+            throw new NotFoundException("Product with id '" + productId + "' not found!");
+        }
     }
 
     @Override
